// src/app/core/models/iuser.ts
import { myUserRole } from '../enums/user-role';

export interface IUser {
  id: string;
  username: string;
  email: string;
  password: string;
  role: myUserRole;
}


// create interface for register
export interface IRegisterUser {
  username: string;
  email: string;
  password: string;
}


// create interface for register
export interface IRegisterResponseUser {
  id: string;
  token: string;
  role: myUserRole;
}


export interface ILoginUser {
  email: string;
  password: string;
}